+++ src/youwol/backends/tree_db/routers/__init__.py
@@ -1,4 +1,5 @@
 # relative
+from .drives import router as router_drives
 from .entities import router as router_entities
 from .folders import router as router_folders
 from .items import router as router_items
